chore: rename to IQ.Wiki
